enable warnings for build-tool and fix warnings on windows

+++ src/platform/win32_string_helpers.c
@@ -14,9 +14,9 @@
 oc_str16 oc_win32_utf8_to_wide(oc_arena* arena, oc_str8 s)
 {
     oc_str16 res = { 0 };
-    res.len = 1 + MultiByteToWideChar(CP_UTF8, 0, s.ptr, s.len, NULL, 0);
+    res.len = 1 + MultiByteToWideChar(CP_UTF8, 0, s.ptr, (int)s.len, NULL, 0);
     res.ptr = oc_arena_push_array(arena, u16, res.len);
-    MultiByteToWideChar(CP_UTF8, 0, s.ptr, s.len, res.ptr, res.len);
+    MultiByteToWideChar(CP_UTF8, 0, s.ptr, (int)s.len, res.ptr, (int)res.len);
     res.ptr[res.len - 1] = '\0';
     return (res);
 }
@@ -24,9 +24,9 @@ oc_str16 oc_win32_utf8_to_wide(oc_arena* arena, oc_str8 s)
 oc_str8 oc_win32_wide_to_utf8(oc_arena* arena, oc_str16 s)
 {
     oc_str8 res = { 0 };
-    res.len = WideCharToMultiByte(CP_UTF8, 0, s.ptr, s.len, NULL, 0, NULL, NULL);
+    res.len = WideCharToMultiByte(CP_UTF8, 0, s.ptr, (int)s.len, NULL, 0, NULL, NULL);
     res.ptr = oc_arena_push_array(arena, u8, res.len + 1);
-    WideCharToMultiByte(CP_UTF8, 0, s.ptr, s.len, res.ptr, res.len, NULL, NULL);
+    WideCharToMultiByte(CP_UTF8, 0, s.ptr, (int)s.len, res.ptr, (int)res.len, NULL, NULL);
     res.ptr[res.len] = '\0';
     return (res);
 }
